requirements.txt
fastapi
uvicorn
pymongo
websockets
python-dotenv
pydantic
